+++ jstests/core/columnstore_large_array_index_correctness.js
@@ -2,15 +2,26 @@
  * Testing of just the query layer's integration for columnar indexes that encode large arrays.
  * @tags: [
  *   # column store indexes are still under a feature flag and require full sbe
- *   uses_column_store_index,
  *   featureFlagColumnstoreIndexes,
  *   featureFlagSbeFull,
+ *   # Columnstore tests set server parameters to disable columnstore query planning heuristics -
+ *   # 1) server parameters are stored in-memory only so are not transferred onto the recipient,
+ *   # 2) server parameters may not be set in stepdown passthroughs because it is a command that may
+ *   #      return different values after a failover
+ *   tenant_migration_incompatible,
+ *   does_not_support_stepdowns,
+ *   not_allowed_with_security_token,
  * ]
  */
 (function() {
 "use strict";
 
-load("jstests/libs/analyze_plan.js");  // For "planHasStage."
+load("jstests/libs/analyze_plan.js");      // For "planHasStage."
+load("jstests/libs/columnstore_util.js");  // For "setUpServerForColumnStoreIndexTest."
+
+if (!setUpServerForColumnStoreIndexTest(db)) {
+    return;
+}
 
 const coll = db.columnstore_large_array_index_correctness;
 coll.drop();
